Refactor MainScreen to support orientation changes and use KioskSettings
The `MainScreen` composable now accepts an `Activity` and a `Modifier` as parameters. It uses `KioskSettingsFactory` to retrieve kiosk settings instead of `DataStoreHelper`.

The `WebViewComponent` is now keyed with `configuration.orientation` to handle orientation changes. It also receives the `activity` and `modifier` from `MainScreen`.

The `Box` composable that previously wrapped `WebViewComponent` and set a black background has been removed.